pyproject.toml
[tool.poetry]
name = "boilerplate_django"
version = "0.1.0"
description = "Plantilla per a nous projectes web amb Django."
authors = ["Codi Cooperatiu <hola@codi.coop>"]
license = "gpl-3.0"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
django-environ = "^0.8.1"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.3"
whitenoise = "^6.0.0"
django-maintenance-mode = "^0.16.3"
tox = "^3.25.0"
django-extensions = "^3.1.5"
django-grappelli = "^3.0.3"
django-constance = {extras = ["database"], version = "^2.8.0"}
django-storages = "^1.12.3"
boto3 = "^1.21.46"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
flake8-bugbear = "^22.3.23"
factory-boy = "^3.2.1"
pytest = "^7.1.1"
pytest-django = "^4.5.2"
pytest-sugar = "^0.9.4"

[tool.black]
exclude = '''
/(
  | migrations
)/

'''

[tool.isort]
skip = "migrations"
multi_line_output = 3
include_trailing_comma = "True"
force_grid_wrap = 0
use_parentheses = "True"
ensure_newline_before_comments = "True"
line_length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
